Update interface, add module system (mostly secure?)
Also remove "Login" folder/files in favor of a JavaScript login integrated into the header (to be implimented later).
The module system is designed to be mostly secure. I plan on doing what SMF does, except with a boolean variable instead of a "define()" to check if the module is being called directly or from index.php.
The root interface (header/sidebar) is pretty much done. Just a few tweaks left after we get the main interfaces done for all of the modules.