Integrate interface files as the structure is pretty much done.
All that's left is to create Modules themselves... I have found this to be very very hard :(
I can't quite figure out how I want the project/user management to work. The best way would to use AJAX but I really don't want to use that.

The Manage_Projects module has some code on it already, but I have a feeling none of it is going to make it to the end product.