Finish Project module interface work and polish module loading script
Ahh, there's a lot to say. I'm proud of what I got done on the add-project page, but without linking to a database, the rest of
what I can do is limited (the modify_project page, for example). I have a few things planed out that don't agree with some code
in this commit, but I'll get that worked out eventually.

Signed-off-by: Daktyl198 <rootme198@outlook.com>