Finish Project View design and switch to pulling jQuery from the CDN
I with the next commit, I plan to have it pulling info from the DB and to depreciate the "Modify Project" submodule.
There is one issue I'm currently working out: the title on the View Project submodule is not selectable without selecing the image with it (which is still hard to do). I think this has
something to do with the fact that the image is using float:left; but I'm not 100% sure...

Signed-off-by: Daktyl198 <rootme198@outlook.com>