Another "I'm not dead" push. SO. BUSY. IRL. Q__Q
I updated the "Member View" page to work with your UserIO.php fix...
I also added an "Edit Project" button on the "Project View" page, but I can't seem to get the input fields to replace the text fields
correctly. We'll have to come back to this, I think. There's a good chance we'll be using full-on JavaScript to do it instead of my original
idea.

Signed-off-by: Daktyl198 <rootme198@outlook.com>